pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentshield-sdk"
version = "0.1.0"
description = "The security primitive the agent ecosystem is missing."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Aditya Belhekar", email = "adityabelhekar@gmail.com" }]

keywords = [
  "ai-agents", "llm-security", "prompt-injection",
  "langchain", "agent-security", "ai-safety",
  "goal-drift", "tool-chain", "runtime-security"
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Typing :: Typed",
]

dependencies = [
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "loguru>=0.7.0",
  "langchain-core>=0.2.0",
  "redis>=5.0.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.4.0",
  "sentence-transformers>=2.6.0",
  "faiss-cpu>=1.8.0",
  "cryptography>=42.0.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
redis = [
  "hiredis>=2.3.0",
]
otel = [
  "opentelemetry-api>=1.24.0",
  "opentelemetry-sdk>=1.24.0",
  "opentelemetry-exporter-otlp>=1.24.0",
]
backend = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.29.0",
  "websockets>=12.0",
]
all = [
  "agentshield-sdk[redis,otel,backend]",
]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "mypy>=1.10.0",
  "ruff>=0.4.0",
  "black>=24.4.0",
  "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://github.com/AdityaBelhekar/AgentShield"
Documentation = "https://AdityaBelhekar.github.io/AgentShield"
Repository = "https://github.com/AdityaBelhekar/AgentShield"
"Bug Tracker" = "https://github.com/AdityaBelhekar/AgentShield/issues"
Changelog = "https://github.com/AdityaBelhekar/AgentShield/blob/main/CHANGELOG.md"

[project.scripts]
agentshield = "agentshield.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["agentshield"]

[tool.hatch.build.targets.sdist]
include = [
  "/agentshield",
  "/README.md",
  "/CHANGELOG.md",
  "/LICENSE",
  "py.typed",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"agentshield/__init__.py" = ["I001"]

[tool.black]
line-length = 100
target-version = ["py311"]
extend-exclude = '''
(
  ^/agentshield/detection/
  | ^/agentshield/interceptors/
  | ^/agentshield/runtime.py
)
'''

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["scratch/tests"]
addopts = "--tb=short -q"

[tool.coverage.run]
source = ["agentshield"]
omit = ["*/scratch/*", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
]
